Fix routed-MLP scatter dtype mismatch under autocast
MoELayer._routed_mlp preallocated its output buffer with torch.zeros_like
of the norm2 output. Under CUDA autocast, LayerNorm is forced to fp32, so
that buffer is fp32 while the expert Linear runs in bf16 — the masked
scatter out[mask] = expert(x[mask]) then raised 'Index put requires source
and destination dtypes match'. This crashed the capacity probe's first
forward on H100 (found by the Modal smoke run). Cast the expert output to
the buffer dtype, matching every other masked-scatter site in the model.

CPU autocast preserves LayerNorm's input dtype, so a full-model CPU
forward can't surface this; the regression test calls _routed_mlp with an
fp32 input under CPU autocast to mirror the CUDA fp32-norm / bf16-expert
split exactly.

+++ src/model/blocks.py
@@ -392,7 +392,11 @@ class MoELayer(nn.Module):
             mask = (expert_ids == e)
             if expert_lens_cpu is None and not mask.any():
                 continue
-            out[mask] = self.routed_mlps[e](x[mask])
+            # `out` inherits x's dtype (fp32, from the fp32 LayerNorm that
+            # produced x under autocast), but the expert Linear runs in the
+            # autocast dtype (bf16) — cast on write so the scatter's src and
+            # dst dtypes match, as every other masked-scatter site does.
+            out[mask] = self.routed_mlps[e](x[mask]).to(out.dtype)
         return out
